pyproject.toml
[project]
name = "devindex-agents"
version = "0.3.0"
description = "DevIndex LangGraph agent for GitHub repository analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "tools", "models", "database", "utils"]


[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
